Add service from the xml export to the database and add foreign key to service from frequency

+++ gofly-batch-partition-airspace-data/src/main/java/fr/gofly/model/Airspace.java
@@ -13,7 +13,7 @@ public class Airspace {
     @Id
     @Column(name = "airspace_id")
     @XmlAttribute(name = "pk")
-    private Long airspaceId;
+    private int airspaceId;
 
     @ManyToOne
     @JoinColumn(name = "territory_id")
